requirements.txt
fastapi>=0.95.0,<0.96.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
pandas>=1.3.0,<1.4.0
openpyxl>=3.0.7,<3.1.0
python-dotenv>=0.19.0,<0.20.0
pydantic>=1.9.0,<2.0.0
openai>=1.0.0
matplotlib>=3.4.0,<3.5.0
plotly>=5.3.0,<5.4.0
pytest>=6.2.5,<6.3.0
pytest-asyncio>=0.15.1,<0.16.0
black>=21.7b0,<22.0
isort>=5.9.3,<5.10.0
flake8>=3.9.2,<3.10.0
psutil>=5.8.0,<5.9.0
httpx>=0.24.0,<0.25.0 